update maths from last commit to oilCalc.py
Add the oilCalv maths

use the fix in the latest version v0.2 of the python oilCalc.py
now gets maths right as working distance definition is correct
and so oil thickness calc is right.

Working distance  for 60x/1.42 is correct,
others are wrong, need to find thm out and fix

also temp and wavelength corrections are still to do
so i disabled the sliders, greyed out. using disable flag
of jquery mobile slider.